Move all the tests over to /src instead
This is so we can test the real code, not the compiled code. The compiled code is supposed to be... uhh... compiled. 

This does mean we're not testing the code that we're actually running, though in practice if the code breaks between writing and compilation then we've done something disastrously wrong.

Also, this doesn't pass the tests because the tests expect certain things from us that we don't do - see following commits.