support switching .start event when they race
when both parties send a .start event after a .ready, the MSC
says that (if the method is the same), we should pick the .start
event if the smallest user. At this point we might have already
returned a verifier to the SDK consumer, so we need to switch to
the new start event internally.