Add some logging around sync for stuck message debugging
For https://github.com/matrix-org/synapse/issues/7206

Intended to be removed (or possibly converted into a proper off-by-default setting) once the issue concludes.

There's a bunch of null checking through defaults here because this is a very critical part of the whole stack. We default to writing errors to the console, but do not break the app for failing to log info about the sync.

This may have adverse affects on performance for large accounts - use `localStorage.setItem("mx_skip_sync_logging", "true")` to disable the worst parts.