Use a temporary Git dependency instead of patch-package
patch-package is only designed for use in top-level applications, not in libraries which get repackaged, and so will fail to run in upstream consumers of matrix-js-sdk. Until the issues with matrix-events-sdk are resolved, this replaces the patch-package workaround with a temporary Git dependency to unblock upstream packages such as Element Call, while keeping the typechecker happy.