Allow read receipt accumulation per room
Following MSC3771, read receipts are not unique per room per user id.
It is now possible for a single user id to have multiple read receipts across threads and the main room timeline.

The sync accumulator was not taking this into consideration and was always persisting the last read receipt it had received